requirements.txt
Django==3.0.6
psycopg2==2.8.5
jsonfield==3.1.0
django-jalali-date==0.3.1